Fjerner funksjonalitet for legacy søk
