+++ src/main/resources/lib/search/search-repo.ts
@@ -1,70 +1,20 @@
 import * as repoLib from '/lib/xp/repo';
 import { RepoConnection } from '/lib/xp/node';
-import { logger } from '../../utils/logging';
-import { runInContext } from '../../context/run-in-context';
-import { SEARCH_REPO_ID } from '../../constants';
-import {
-    SEARCH_REPO_CONFIG_NODE_LEGACY,
-    SEARCH_REPO_CONTENT_BASE_NODE,
-    SEARCH_REPO_DELETION_QUEUE_BASE_NODE,
-    SEARCH_REPO_UPDATE_STATE_NODE,
-} from './search-utils';
-import { forceArray } from '../../utils/array-utils';
-import { getSearchRepoConnection, SEARCH_REPO_CONFIG_NODE } from '../utils';
+import { logger } from '../utils/logging';
+import { runInContext } from '../context/run-in-context';
+import { SEARCH_REPO_ID } from '../constants';
+import { getSearchRepoConnection, SEARCH_REPO_CONFIG_NODE } from './utils';
+
+const SEARCH_REPO_DELETION_QUEUE_BASE_NODE = 'deletionQueue';
+const SEARCH_REPO_CONTENT_BASE_NODE = 'content';
+const SEARCH_REPO_UPDATE_STATE_NODE = 'updateState';
+const SEARCH_REPO_CONFIG_NODE_LEGACY = 'config';
 
 const BASE_NODE_KEY = `/${SEARCH_REPO_CONTENT_BASE_NODE}`;
 const DELETION_QUEUE_NODE_KEY = `/${SEARCH_REPO_DELETION_QUEUE_BASE_NODE}`;
 const UPDATE_STATE_NODE_KEY = `/${SEARCH_REPO_UPDATE_STATE_NODE}`;
 const CONFIG_NODE_KEY = `/${SEARCH_REPO_CONFIG_NODE_LEGACY}`;
 const CONFIG_EXTERNAL_NODE_KEY = `/${SEARCH_REPO_CONFIG_NODE}`;
-
-type UpdateQueueEntry = {
-    contentId: string;
-    repoId: string;
-};
-
-type UpdateQueueState = {
-    isQueuedUpdateAll?: boolean;
-    queuedContentUpdates?: UpdateQueueEntry[];
-};
-
-export const getUpdateQueue = () => {
-    const repo = getSearchRepoConnection();
-    return repo.get<UpdateQueueState>(UPDATE_STATE_NODE_KEY);
-};
-
-const setUpdateState = (state: UpdateQueueState) => {
-    const repo = getSearchRepoConnection();
-    repo.modify({
-        key: UPDATE_STATE_NODE_KEY,
-        editor: (node) => ({
-            ...node,
-            ...state,
-        }),
-    });
-};
-
-export const queueUpdateAll = () => {
-    setUpdateState({ isQueuedUpdateAll: true, queuedContentUpdates: [] });
-};
-
-export const queueUpdateForContent = (contentId: string, repoId: string) => {
-    const currentUpdateState = getUpdateQueue();
-    const newEntry = { contentId, repoId };
-
-    setUpdateState({
-        queuedContentUpdates: [
-            ...(currentUpdateState?.queuedContentUpdates
-                ? forceArray(currentUpdateState.queuedContentUpdates)
-                : []),
-            newEntry,
-        ],
-    });
-};
-
-export const clearSearchNodeUpdateQueue = () => {
-    setUpdateState({ isQueuedUpdateAll: false, queuedContentUpdates: [] });
-};
 
 const getSearchRepo = () => {
     try {
